Fix table pagination issue
