+++ GUI/src/components/DataTable/index.tsx
@@ -1,4 +1,4 @@
-import React, { CSSProperties, FC, ReactNode, useId, useState } from 'react'
+import React, { CSSProperties, FC, ReactNode, useEffect, useState } from 'react';
 import {
   ColumnDef,
   useReactTable,
@@ -14,17 +14,12 @@ import {
   Row,
   RowData,
   ColumnFiltersState,
-  PaginationState,
-} from '@tanstack/react-table'
-import { RankingInfo, rankItem } from '@tanstack/match-sorter-utils'
-import { MdUnfoldMore, MdExpandMore, MdExpandLess, MdOutlineEast, MdOutlineWest } from 'react-icons/md'
-import clsx from 'clsx'
-import { Link } from 'react-router-dom'
-import { useTranslation } from 'react-i18next'
-
-import { Icon, Track } from '../../components'
-import Filter from './Filter'
-import './DataTable.scss'
+} from '@tanstack/react-table';
+import { RankingInfo, rankItem } from '@tanstack/match-sorter-utils';
+import TableHeader from './TableHeader';
+import TableFooter, { pageSizeOptions } from './TableFooter';
+import './DataTable.scss';
+import { useSearchParams } from 'react-router-dom';
 
 type DataTableProps = {
   data: any
@@ -38,12 +33,6 @@ type DataTableProps = {
   setColumnVisibility?: React.Dispatch<React.SetStateAction<VisibilityState>>
   disableHead?: boolean
   meta?: TableMeta<any>
-}
-
-type ColumnMeta = {
-  meta: {
-    size: number | string
-  }
 }
 
 declare module '@tanstack/table-core' {
@@ -61,8 +50,6 @@ declare module '@tanstack/react-table' {
     getRowStyles: (row: Row<TData>) => CSSProperties
   }
 }
-
-type CustomColumnDef = ColumnDef<any> & ColumnMeta
 
 const fuzzyFilter: FilterFn<any> = (row, columnId, value, addMeta) => {
   const itemRank = rankItem(row.getValue(columnId), value)
@@ -85,15 +72,17 @@ const DataTable: FC<DataTableProps> = ({
   disableHead,
   meta,
 }) => {
-  const pagesShown = 7;
-  const id = useId()
-  const { t } = useTranslation()
   const [sorting, setSorting] = useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>([])
   const [pageIndex, setPageIndex] = useState(0);
-  const [pageSize, setPageSize] = useState(10);
-  const pagination = { pageIndex, pageSize };
+  const [pageSize, setPageSize] = useState(pageSizeOptions[0]);
+  const [params] = useSearchParams();
+  const page = params.get('page');
   
+  useEffect(() => {
+    setPageIndex(page ? Number(page) - 1 : 0);
+  }, [page]);
+
   const table = useReactTable({
     data,
     columns,
@@ -105,7 +94,10 @@ const DataTable: FC<DataTableProps> = ({
       columnFilters,
       globalFilter,
       columnVisibility,
-      ...{ pagination },
+      pagination: { 
+        pageIndex,
+        pageSize,
+      },
     },
     meta,
     onColumnFiltersChange: setColumnFilters,
@@ -113,67 +105,16 @@ const DataTable: FC<DataTableProps> = ({
     onColumnVisibilityChange: setColumnVisibility,
     globalFilterFn: fuzzyFilter,
     onSortingChange: setSorting,
-    onPaginationChange: () => (prev: PaginationState) => {
-      setPageIndex(prev.pageIndex);
-      setPageSize(prev.pageSize);
-    },
     getCoreRowModel: getCoreRowModel(),
     getFilteredRowModel: getFilteredRowModel(),
-    ...(pagination && { getPaginationRowModel: getPaginationRowModel() }),
+    getPaginationRowModel: getPaginationRowModel(),
     ...(sortable && { getSortedRowModel: getSortedRowModel() }),
   })
-
-  const getPages = (): number[] => {
-    const current = table.getState().pagination.pageIndex;
-    const pages: number[] = [];
-    const lastShownPage = Math.min(table.getPageCount(), current + pagesShown);
-
-    pages.push(current);
-    for (; pages.length < Math.min(pagesShown, table.getPageCount());) {
-      if (pages[0] > 0) pages.unshift(pages[0] - 1);
-      if (pages[pages.length - 1] + 1 < lastShownPage) {
-        pages.push(pages[pages.length - 1] + 1);
-      }
-    }
-
-    return pages;
-  };
 
   return (
     <div className="data-table__wrapper">
       <table className="data-table">
-        {!disableHead && (
-          <thead>
-            {table.getHeaderGroups().map((headerGroup) => (
-              <tr key={headerGroup.id}>
-                {headerGroup.headers.map((header) => (
-                  <th key={header.id} style={{ width: (header.column.columnDef as CustomColumnDef).meta?.size }}>
-                    {header.isPlaceholder ? null : (
-                      <>
-                        <Track gap={8}>
-                          {sortable && header.column.getCanSort() && (
-                            <button onClick={header.column.getToggleSortingHandler()}>
-                              {{
-                                asc: <Icon icon={<MdExpandMore fontSize={20} />} size="medium" />,
-                                desc: <Icon icon={<MdExpandLess fontSize={20} />} size="medium" />,
-                              }[header.column.getIsSorted() as string] ?? (
-                                <Icon icon={<MdUnfoldMore fontSize={22} />} size="medium" />
-                              )}
-                            </button>
-                          )}
-                          {flexRender(header.column.columnDef.header, header.getContext())}
-                          {filterable && header.column.getCanFilter() && (
-                            <Filter column={header.column} table={table} />
-                          )}
-                        </Track>
-                      </>
-                    )}
-                  </th>
-                ))}
-              </tr>
-            ))}
-          </thead>
-        )}
+        {!disableHead && <TableHeader table={table} sortable={sortable} filterable={filterable} />}
         <tbody>
           {tableBodyPrefix}
           {table.getRowModel().rows.map((row) => (
@@ -185,69 +126,7 @@ const DataTable: FC<DataTableProps> = ({
           ))}
         </tbody>
       </table>
-      {pagination && (
-        <div className="data-table__pagination-wrapper">
-          {table.getPageCount() * table.getState().pagination.pageSize > table.getState().pagination.pageSize && (
-            <div className="data-table__pagination">
-              <button className="previous" onClick={() => table.previousPage()} disabled={!table.getCanPreviousPage()}>
-                <MdOutlineWest />
-              </button>
-              <nav role="navigation" aria-label={t('global.paginationNavigation') || ''}>
-                <ul className="links">
-                {getPages().map((page, i) => {
-                    if (
-                      (i === 0 && page !== 0) ||
-                      (i === pagesShown - 1 && page !== table.getPageCount() - 1)
-                    ) {
-                      return <p key={`${id}-${page}`}>...</p>;
-                    }
-                    return (
-                      <li
-                      key={`${id}-${page}`}
-                      className={clsx({ active: table.getState().pagination.pageIndex === page })}
-                    >
-                      <Link
-                        to={`?page=${page + 1}`}
-                        onClick={() => table.setPageIndex(page)}
-                        aria-label={t('global.gotoPage') + page}
-                        aria-current={table.getState().pagination.pageIndex === page}
-                      >
-                        {page + 1}
-                      </Link>
-                    </li>
-                    );
-                  })}
-                </ul>
-              </nav>
-              <button
-                className="next"
-                onClick={() => {
-                  table.nextPage()
-                }}
-                disabled={!table.getCanNextPage()}
-              >
-                <MdOutlineEast />
-              </button>
-            </div>
-          )}
-          <div className="data-table__page-size">
-            <label htmlFor={id}>{t('global.resultCount')}</label>
-            <select
-              id={id}
-              value={table.getState().pagination.pageSize}
-              onChange={(e) => {
-                table.setPageSize(Number(e.target.value))
-              }}
-            >
-              {[10, 20, 30, 40, 50].map((pageSize) => (
-                <option key={pageSize} value={pageSize}>
-                  {pageSize}
-                </option>
-              ))}
-            </select>
-          </div>
-        </div>
-      )}
+      <TableFooter table={table} setPageIndex={setPageIndex} setPageSize={setPageSize} />
     </div>
   )
 }
